@@ -4,15 +4,15 @@
     "license": "MIT",
 
     "require": {
-        "PHP":                  "^7.0.0",
-        "symfony/console":      "~3.2.2",
-        "ultra-lite/container": "~3.1.0"
+        "PHP":                  "^7.0",
+        "symfony/console":      "^3.3",
+        "ultra-lite/container": "^6.1"
     },
 
     "require-dev": {
-        "phpunit/phpunit": "~5.7.9",
-        "phpspec/phpspec": "~3.2.3",
-        "behat/behat":     "~3.3.0"
+        "phpunit/phpunit": "^6.1",
+        "phpspec/phpspec": "^3.4",
+        "behat/behat":     "^3.3"
     },
 
     "autoload": {
